requirements.txt
langchain_openai
langchain_core
langchain_community
python-dotenv
streamlit
langchain_google_genai